Split `replication_messages` into submodules
No logical changes.
